+++ template-unitybepinex/index.js
@@ -205,7 +205,7 @@ const tools = [
     exclusive: true,
     shell: true,
     //defaultPrimary: true,
-    //parameters: PARAMETERS
+    parameters: PARAMETERS
   }, //*/
   {
     id: `${GAME_ID}-customlaunchalt`,
@@ -218,7 +218,7 @@ const tools = [
     exclusive: true,
     shell: true,
     //defaultPrimary: true,
-    //parameters: PARAMETERS
+    parameters: PARAMETERS
   }, //*/
 ];