+++ template-unitymelonloaderbepinex-hybrid/index.js
@@ -436,7 +436,7 @@ const tools = [
     exclusive: true,
     shell: true,
     //defaultPrimary: true,
-    //parameters: PARAMETERS
+    parameters: PARAMETERS,
   }, //*/
   {
       id: SAVEEDITOR_ID,