+++ game-warhammer40kroguetrader/CHANGELOG.md
@@ -4,13 +4,17 @@
 
 - None Planned
 
-## [0.1.1] - 2025-01-16
+## [0.1.2] - 2026-01-21
+
+- Fixed: UMM plugin mod installer could use wrong plugin name for folder (i.e. "0Harmony")
+
+## [0.1.1] - 2026-01-16
 
 - Fixed: Owlcat mod installer removing "Bundles" folder and files as they have no file extension
 - Added: Button to open the "OwlcatModificationManagerSettings.json" file
 - Added: Notification for Xbox users that custom Portaits mods will not be loaded by that version of the game
 - Added: Support for Portrait Manager (installer and launch tool)
 
-## [0.1.0] - 2025-01-13
+## [0.1.0] - 2026-01-13
 
 - Inital Release